Рефакторинг страниц пользователей
