+++ packages/frontend/src/pages/users/index.tsx
@@ -1,23 +1,8 @@
 import type { NextPage } from "next";
-import { useGetUsersList } from "src/shared/api/users/useGetUsersList";
-import Link from "next/link";
+import UsersPage from "src/pagesFsd/users/Users";
 
 const Users: NextPage = () => {
-  const { data } = useGetUsersList();
-  if (!data) {
-    return <>Loading...</>;
-  }
-  return (
-    <div>
-      <ul>
-        {data.map((item) => (
-          <li key={item.id}>
-            <Link href={`/users/${item.id}`}>{`${item.firstName} ${item.lastName}`}</Link>
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
+  return <UsersPage />;
 };
 
 export default Users;
