+++ packages/frontend/src/shared/helpers/isExist.ts
@@ -1,0 +1,2 @@
+export const isExist = <T>(data: T | null | undefined): data is T =>
+  data !== undefined && data !== null;
